grab history to base
